# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResStudent": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"class": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"class",
					"age",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetSTUDENT": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqSTUDENT": {
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"class": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"class",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqSTUDENT": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"class": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"class",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"ISTUDENT": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteSTUDENT": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResTeacher": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"qualification": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"qualification",
					"age",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetTEACHER": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Tokken": {
				"properties": {
					"token": {}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginteacher": {
				"properties": {
					"FirstName": {
						"type": "string"
					}
				},
				"required": [
					"FirstName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqTEACHER": {
				"properties": {
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"qualification": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"FirstName",
					"LastName",
					"qualification",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqTEACHER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"LastName": {
						"type": "string"
					},
					"qualification": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"FirstName",
					"LastName",
					"qualification",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ITEACHER": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "swag",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/student/getstudent": {
			"post": {
				"operationId": "Getstudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResStudent"
								}
							}
						}
					}
				},
				"tags": [
					"student"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetSTUDENT"
							}
						}
					}
				}
			}
		},
		"/student/savestudent": {
			"post": {
				"operationId": "Savestudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResStudent"
								}
							}
						}
					}
				},
				"tags": [
					"student"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqSTUDENT"
							}
						}
					}
				}
			}
		},
		"/student/updatestudent": {
			"put": {
				"operationId": "Updatestudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResStudent"
								}
							}
						}
					}
				},
				"tags": [
					"student"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqSTUDENT"
							}
						}
					}
				}
			}
		},
		"/student/deletestudent": {
			"delete": {
				"operationId": "Deletstudent",
				"responses": {
					"200": {
						"description": "Admin deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISTUDENT"
								}
							}
						}
					}
				},
				"tags": [
					"student"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteSTUDENT"
							}
						}
					}
				}
			}
		},
		"/student/getstudentlist": {
			"post": {
				"operationId": "GetstudentList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResStudent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"student"
				],
				"security": [],
				"parameters": []
			}
		},
		"/teacher/getteacher": {
			"post": {
				"operationId": "Getteacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResTeacher"
								}
							}
						}
					}
				},
				"tags": [
					"teacher"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetTEACHER"
							}
						}
					}
				}
			}
		},
		"/teacher/loginteacher": {
			"post": {
				"operationId": "Loginteacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tokken"
								}
							}
						}
					}
				},
				"tags": [
					"teacher"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginteacher"
							}
						}
					}
				}
			}
		},
		"/teacher/saveteacher": {
			"post": {
				"operationId": "Saveteacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResTeacher"
								}
							}
						}
					}
				},
				"tags": [
					"teacher"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqTEACHER"
							}
						}
					}
				}
			}
		},
		"/teacher/updateteacher": {
			"put": {
				"operationId": "Updateteacher",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResTeacher"
								}
							}
						}
					}
				},
				"tags": [
					"teacher"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqTEACHER"
							}
						}
					}
				}
			}
		},
		"/teacher/deleteteacher": {
			"delete": {
				"operationId": "Deletteacher",
				"responses": {
					"200": {
						"description": "Admin deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITEACHER"
								}
							}
						}
					}
				},
				"tags": [
					"teacher"
				],
				"security": [],
				"parameters": []
			}
		},
		"/teacher/getteacherlist": {
			"post": {
				"operationId": "GetteacherList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResTeacher"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"teacher"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}